requirements.txt
python>=3.7
opencv-python
numpy
pillow
